+++ Changelog.md
@@ -6,9 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## TODO
+## [0.1.3] - Thu Jan 23 11:51:22 AM PST 2025
 
-- Generate test data
+### Added
+
+- Downloaded Llama-8B model from Huggingface
+- Fine tuned model ( Time taken: 0:51:50 )
+
 
 ## [0.1.2] - Thu Jan 23 11:51:22 AM PST 2025